refactor: (#747) 공지사항 쿼리, 커맨드 분리
